+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - User option to disable to-do checkboxes on tabs ([#6](https://github.com/benji300/joplin-note-tabs/issues/6))
 - User option to automatically pin notes when edited ([#11](https://github.com/benji300/joplin-note-tabs/issues/11))
 - Commands to switch between tabs ([#5](https://github.com/benji300/joplin-note-tabs/issues/5))
-- Command to switch to last active note ([#8](https://github.com/benji300/joplin-note-tabs/issues/8))
+- Command to switch to last active tab ([#8](https://github.com/benji300/joplin-note-tabs/issues/8))
 - Ability to pin note(s) from note list context menu ([#12](https://github.com/benji300/joplin-note-tabs/pull/12)
 - Ability to pin note from editor context menu ([#9](https://github.com/benji300/joplin-note-tabs/pull/9) by [@ambrt](https://github.com/ambrt))
 - Ability to pin note with double click on tab